// dub-packages/prisma-oppulence/generated/zod/outputTypeSchemas/LinkFindManyArgsSchema.ts
export const LinkSelectSchema: z.ZodType<Prisma.LinkSelect> = z.object({
  id: z.boolean().optional(),
  domain: z.boolean().optional(),
  key: z.boolean().optional(),
  url: z.boolean().optional(),
  shortLink: z.boolean().optional(),
  archived: z.boolean().optional(),
  expiresAt: z.boolean().optional(),
  expiredUrl: z.boolean().optional(),
  password: z.boolean().optional(),
  trackConversion: z.boolean().optional(),
  proxy: z.boolean().optional(),
  title: z.boolean().optional(),
  description: z.boolean().optional(),
  image: z.boolean().optional(),
  video: z.boolean().optional(),
  utm_source: z.boolean().optional(),
  utm_medium: z.boolean().optional(),
  utm_campaign: z.boolean().optional(),
  utm_term: z.boolean().optional(),
  utm_content: z.boolean().optional(),
  rewrite: z.boolean().optional(),
  doIndex: z.boolean().optional(),
  ios: z.boolean().optional(),
  android: z.boolean().optional(),
  geo: z.boolean().optional(),
  userId: z.boolean().optional(),
  projectId: z.boolean().optional(),
  externalId: z.boolean().optional(),
  tenantId: z.boolean().optional(),
  publicStats: z.boolean().optional(),
  clicks: z.boolean().optional(),
  lastClicked: z.boolean().optional(),
  leads: z.boolean().optional(),
  sales: z.boolean().optional(),
  saleAmount: z.boolean().optional(),
  createdAt: z.boolean().optional(),
  updatedAt: z.boolean().optional(),
  comments: z.boolean().optional(),
  programId: z.boolean().optional(),
  user: z.union([z.boolean(),z.lazy(() => UserArgsSchema)]).optional(),
  project: z.union([z.boolean(),z.lazy(() => ProjectArgsSchema)]).optional(),
  projectDomain: z.union([z.boolean(),z.lazy(() => DomainArgsSchema)]).optional(),
  tags: z.union([z.boolean(),z.lazy(() => LinkTagFindManyArgsSchema)]).optional(),
  webhooks: z.union([z.boolean(),z.lazy(() => LinkWebhookFindManyArgsSchema)]).optional(),
  dashboard: z.union([z.boolean(),z.lazy(() => DashboardArgsSchema)]).optional(),
  programEnrollment: z.union([z.boolean(),z.lazy(() => ProgramEnrollmentArgsSchema)]).optional(),
  programInvite: z.union([z.boolean(),z.lazy(() => ProgramInviteArgsSchema)]).optional(),
  program: z.union([z.boolean(),z.lazy(() => ProgramArgsSchema)]).optional(),
  customers: z.union([z.boolean(),z.lazy(() => CustomerFindManyArgsSchema)]).optional(),
  _count: z.union([z.boolean(),z.lazy(() => LinkCountOutputTypeArgsSchema)]).optional(),
}).strict()

export const LinkFindManyArgsSchema: z.ZodType<Prisma.LinkFindManyArgs> = z.object({
  select: LinkSelectSchema.optional(),
  include: LinkIncludeSchema.optional(),
  where: LinkWhereInputSchema.optional(),
  orderBy: z.union([ LinkOrderByWithRelationInputSchema.array(),LinkOrderByWithRelationInputSchema ]).optional(),
  cursor: LinkWhereUniqueInputSchema.optional(),
  take: z.number().optional(),
  skip: z.number().optional(),
  distinct: z.union([ LinkScalarFieldEnumSchema,LinkScalarFieldEnumSchema.array() ]).optional(),
}).strict() ;